backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
python-multipart
boto3
python-jose[cryptography]
python-magic
aiofiles
